DataHelper object for ctxcontent is extracted.
fix: plugin loading from DOMContentLoaded event is fixed.
