+++ src/main.js
@@ -1,5 +1,7 @@
-(async () => {
+async function init(ev) {
   const src = chrome.runtime.getURL("./ctxcontent/index.js");
   const contentScript = await import(src);
   contentScript.main();
-})();
+};
+
+window.addEventListener("DOMContentLoaded", init);
